Add files via upload
This Java code implements an Employee Management System (EMS), which is a software system designed to handle various tasks related to employee information and management within an organization. The EMS program is divided into multiple classes, each responsible for specific functionalities.

MainMenu Class: This class presents the main menu of the EMS to the user. It displays options such as adding employee details, viewing employee details, removing employees, updating employee details, and exiting the program. This class aims to provide a clear and user-friendly interface for navigating through the EMS functionalities.

EmployDetail Class: This class represents the structure of an employee's details. It prompts the user to input essential information about an employee, such as their name, father's name, ID, contact details, position, and salary. The collected information is intended for further processing and storage.

Employee_Add Class: This class handles the addition of employee details. It utilizes the information obtained from the EmployDetail class to create a text file named after the employee's ID. The file contains the collected employee information. The class ensures that duplicate employees are not added by checking if the file already exists.

Employee_Show Class: This class facilitates the display of employee details. It reads the contents of a designated text file associated with an employee's ID and presents the stored information to the user. This functionality aids in viewing employee details efficiently.

Employee_Remove Class: This class deals with the removal of employee records. By referencing the employee's ID, it identifies and deletes the corresponding text file containing their details. The class performs verification to determine if the employee exists before proceeding with the removal.

Employee_Update Class: This class enables the modification of employee details. It offers the user the ability to update specific information by identifying the employee using their ID. The class reads the existing details, allows the user to specify the information to be updated, and then replaces the old data with the new input.

CodeExit Class: This class handles the graceful termination of the EMS program. It displays a closing message to the user and calls the System.exit(0) method to exit the program with a status code indicating successful termination.

javaEmployeeManagement Class (Main): This is the main class that orchestrates the entire EMS program. It initiates the program by displaying the main menu using the MainMenu class. It then enters a loop to process user input and execute the corresponding actions based on their choices. The loop continues until the user decides to exit the program.